Fix stock initialization
Fixed calculation of intialization period to fix jump in stock additions in 2021. Also use statistical stock data for 2000 and 2015 values of VEH_STCK_TOT, and linear regression between these for all other years, including the optimization period. This is done in the intialization period as well to get a smooth curve