Fix infeasibility issue for full "real" dataset
Technology addition constraint (EQ_STCK_GRD) was defined for all tec, but the incumbent technology should be unconstrained, as model will be infeasible if total additions to stock exceed the add gradient (constraint) value.
To solve the issue, newtecs, a subset of emerging technologies (i.e., not incumbent) was introduced up for this constraint.